Code/requirements.txt
asgiref==3.2.10
cffi==1.14.2
cryptography==3.1
Django==3.1.1
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
func-timeout==4.3.5
joblib==0.16.0
nameparser==1.0.6
numpy==1.19.2
pandas==1.1.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.22.1
scipy==1.5.2
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
